Move ClamAv status into separate namespace
